[!] Finally fixed manually created makefiles for linux (not cmake) and run Simulator with GUI on Ubuntu 20.04. Deploy missed files Qt6XcbQpa.* and Qt6OpenGL.*. Unused variables warnings treated as errors also were fixed. [*] Next step to build cmake generated makefiles of linux

+++ debugger/src/cpu_fnc_plugin/cpu_riscv_func.cpp
@@ -124,9 +124,9 @@ void CpuRiver_Functional::handleTrap() {
     }
 
     int pendidx = -1;
-    for (int i = 0; i < 8*sizeof(interrupt_pending_); i++) {
+    for (size_t i = 0; i < 8*sizeof(interrupt_pending_); i++) {
         if (interrupt_pending_[i >> 6] & (1ull << (i & 0x3f))) {
-            pendidx = i;
+            pendidx = static_cast<int>(i);
             break;
         }
     }
